requirements.txt
tensorflow>=2.15.0
keras-tuner>=1.4.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
yfinance>=0.2.0
fredapi>=0.5.0
scipy>=1.11.0
python-dotenv>=1.0.0
statsmodels>=0.14.0